README.md
This code is to run the Arduino Uno and it is written using C language. The main concept of this project is to design a circuit which  will detect the distance of the object and displays on the 
analog meter. The sensory output from the ultrasonic 
module is used to trigger the rotation of the servo motor. 
The degree of rotation of the motor is directly proportional 
to the distance of the object from the ultrasonic module.